% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{brief_item}
\alias{brief_item}
\title{Make a Latex briefitem}
\usage{
brief_item(what = "", when = "", where = "")
}
\arguments{
\item{what}{Param what.}

\item{when}{Param when.}

\item{where}{Param where.}
}
\value{
A chracter vector.
}
\description{
Make a Latex briefitem
}
\examples{
brief_item(what = "Hallo", when = "world", where = "here")
}
