% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{brief_section}
\alias{brief_section}
\title{Make a Latex briefsection}
\usage{
brief_section(what = " ", when = " ", where = " ")
}
\arguments{
\item{what}{Param what.}

\item{when}{Param when.}

\item{where}{Param where.}
}
\value{
A chracter vector.
}
\description{
Make a Latex briefsection
}
\examples{
brief_section(what = "Hallo", when = "world", where = "here")
}
