% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{briefitem}
\alias{briefitem}
\title{Make a briefitem}
\usage{
briefitem(what = "", when = "", where = "")
}
\arguments{
\item{what}{Param.}

\item{when}{Param.}

\item{where}{Param.}
}
\value{
A string
}
\description{
Make a briefitem
}
\examples{
briefitem(what = "Hallo", when = "world", where = "here")
}
