% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{briefsection}
\alias{briefsection}
\title{Make a briefsection}
\usage{
briefsection(what = " ", when = " ", where = " ")
}
\arguments{
\item{what}{Param.}

\item{when}{Param.}

\item{where}{Param.}
}
\value{
A string
}
\description{
Make a briefsection
}
\examples{
briefsection(what = "Hallo", when = "world", where = "here")
}
