% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitfun.R
\name{copycat_gitsearch}
\alias{copycat_gitsearch}
\title{copycat_gitsearch}
\usage{
copycat_gitsearch(author, repository, branch = "master")
}
\arguments{
\item{author}{A character vector.}

\item{repository}{A character vector.}

\item{branch}{A character vector.}
}
\value{
A character vector.
}
\description{
copycat_gitsearch
}
