% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{cventry}
\alias{cventry}
\title{Make a cventry entry}
\usage{
cventry(
  center = " ",
  lefttop = " ",
  righttop = " ",
  rightbottom = " ",
  leftbuttom = " "
)
}
\arguments{
\item{center}{Param.}

\item{lefttop}{Param.}

\item{righttop}{Param.}

\item{rightbottom}{Param.}

\item{leftbuttom}{Param.}
}
\value{
A ggplot2 object.
}
\description{
Make a cventry entry
}
\examples{
cventry(center = "Hallo", lefttop = "world")

}
