% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{edgar_packages}
\alias{edgar_packages}
\title{List all packages imported by the edgar package}
\usage{
edgar_packages()
}
\description{
List all packages imported by the edgar package
}
\examples{
edgar_packages()
}
