% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilingHeader.R
\name{extractState}
\alias{extractState}
\title{Extract State abbreviation from free text}
\usage{
extractState(state_text, marker)
}
\arguments{
\item{state_text}{text in which a marker for state (e.g. 'state of incorporation')}

\item{marker}{is the state marker to detect}
}
\value{
state abbreviation
}
\description{
This function reads and parses state abbreviation from SEC Edgar
filings
}
