% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilingHeader.R
\name{GetConfirmedPeriodOfReport}
\alias{GetConfirmedPeriodOfReport}
\title{Scrape the reporting period of the header from the filing text}
\usage{
GetConfirmedPeriodOfReport(header)
}
\arguments{
\item{header}{the filing text}
}
\value{
Returns that period that the filing covers
}
\description{
Scrape the reporting period of the header from the filing text
}
