% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDailyMaster.R
\name{getDailyMaster}
\alias{getDailyMaster}
\title{Retrieves daily master index.}
\usage{
getDailyMaster(input.date)
}
\arguments{
\item{input.date}{in character format 'mm/dd/YYYY'.}
}
\value{
Function returns filing information for the input date.
}
\description{
\code{getDailyMaster} retrieves daily master index from US SEC site.
}
\details{
getDailyMaster function takes date as an input parameter from user,  
download master index for that particular date from https://www.sec.gov/Archives/edgar/daily-index/
site. It strips headers and convert this daily filing information in data frame format.
Function creates new directory 'Daily Index' into working directory 
to save these downloaded daily master index files.
}
\examples{
\dontrun{

daily.filing.info <- getDailyMaster('08/09/2016')
} 
}

