% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilingHeaderForm13.R
\name{GetFilingCusip}
\alias{GetFilingCusip}
\title{Scrape the CUSIP from the filing text}
\usage{
GetFilingCusip(filing.text)
}
\arguments{
\item{filing.text}{the filing text}
}
\value{
Returns CUSIP of the Subject that the filing covers
}
\description{
Scrape the CUSIP from the filing text
}
