% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilingHeaderForm13.R
\name{GetFilingEventDate}
\alias{GetFilingEventDate}
\title{Scrape the event date from the filing text}
\usage{
GetFilingEventDate(filing.text)
}
\arguments{
\item{filing.text}{the filing text}
}
\value{
Returns date of the event that the filing covers
}
\description{
Scrape the event date from the filing text
}
