% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilingHeaderForm13.R
\name{getFilingHeaderForm13}
\alias{getFilingHeaderForm13}
\title{Scrape EDGAR filing header information for From 13G/D}
\usage{
getFilingHeaderForm13(cik.no, form.type = c("SC 13G", "SC 13D"), 
filing.year, useragent, ...)
}
\arguments{
\item{cik.no}{vector of CIK(s) in integer format. Suppress leading
zeroes from CIKs. cik.no = 'ALL' conisders all the CIKs.}

\item{form.type}{character vector containing form type to be downloaded.
form.type = c("SC 13G", "SC 13D") is the default.}

\item{filing.year}{vector of four digit numeric year}

\item{useragent}{Should be in the form of "Your Name Contact@domain.com"}

\item{...}{for passing arguments between functions}
}
\value{
Function returns dataframe containing the CIK number, company name,
date of filing, accession number, date of event, fiscal year end,
Standard Industrial Classification (SIC) code, Internal Revenue Code (IRS)
code, state of incorporation, business address, and mailing address.
The form type 13 contain multiple filers (i.e. the subject and the filer) hence the
output contains header information on both of them.
}
\description{
\code{getFilingHeaderForm13} Extract EDGAR filing header information for From 13G/D
}
\details{
getFilingHeaderForm13 function takes CIK(s), form types (SC 13G or SC 13D), 
and year(s) as input parameters. The function first imports available 
downloaded filings in local working directory Edgar filings_full text created by 
\link[edgar:getFilings]{getFilings} function; otherwise, it automatically downloads the 
filings which are not already been downloaded. It then parses all the important 
header information from filings. The function returns a data.frame with filing and 
header information. According to SEC EDGAR's guidelines a user also needs to 
declare user agent. The function uses the \link[future.apply:future_lapply]{future_lapply} 
function to extract information in parallel. Please set \link[future:plan]{plan} 
according to your needs before running the function. The progress bar
can be controlled with the \link[progressr]{progressr} package. See in particular 
\link[progressr:handlers]{handlers}

The CUSIP number is for the subject (i.e. the same company for which the CIK was supplied).
To distinguish between subject and filer look at the filer.type column.
The subject.cusips are extracted as they are presented in the
filing. Some post-processing may be necessary. For instance, cusips for the same company sometimes have
varying lengths. When the period.of.report is NA it can often be set to the last day of the year.
}
\examples{
\dontrun{
library(future.apply)
plan(multisession, workers = availableCores()-1)

# if a progress update is desired
library(progressr)
handlers(global = TRUE)

header.df <- getFilingHeaderForm13(cik.no = c('1000180', '38079'),
                         form.type = 'SC 13G', filing.year = 2006, useragent)

header.df <- getFilingHeaderForm13(cik.no = '1000180', 'SC 13D',
                         filing.year = c(2005, 2006), useragent)
}
}
