% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilingInfo.R
\name{getFilingInfo}
\alias{getFilingInfo}
\title{Retrieves filing information of company.}
\usage{
getFilingInfo(firm.name, filing.year)
}
\arguments{
\item{firm.name}{Desired full name or partial name of company.}

\item{filing.year}{Filing year in integer.}
}
\value{
Function returns dataframe with filing information.
}
\description{
\code{getFilingInfo} retrieves filing information of company based on company name.
}
\details{
getFilingInfo function takes company name and filing year as input parameters from user,
and gives filing information for that company. Master index file must be present 
for given year before applying this function which can be downloaded using 
\link[edgar]{getMasterIndex} function.
}
\examples{
\dontrun{

info <- getFilingInfo('United Technologies', 1994) 
}
}
