% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_wordfrquency.R
\name{GetWordfrquency}
\alias{GetWordfrquency}
\title{Creates words frequency dataframe of 10-K statement.}
\usage{
GetWordfrquency()
}
\value{
Function returns words frequency dataframe.
}
\description{
\code{GetWordfrquency} creates word frequency dataframe of 10-K statement.
}
\details{
GetWordfrquency function asks the user to locate 10-K statement which can be downloaded using
\link[edgar]{Downloadfilings} function. Function cleans text of 10-K statement
and creates words frequency dataframe. This words frequency dataframe is used
in the functions \link[edgar]{PositiveWordcloud} and \link[edgar]{NegativeWordcloud}.
}
\examples{
\dontrun{

word.frq <- GetWordfrquency()
}
}

