% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggpie}
\alias{ggpie}
\title{Make a pie chart.}
\usage{
ggpie(data, fill, legend = "none")
}
\arguments{
\item{data}{Data.}

\item{fill}{The variable to fill the pie.}

\item{legend}{The legend position.}
}
\value{
A ggplot2 object.
}
\description{
Make a pie chart.
}
\examples{
ggpie(palmerpenguins::penguins, fill = island)
}
