\name{LMMasterDictionary}
\alias{LMMasterDictionary}
\title{Loughran and McDonald Sentiment Master Dictionary}
\source{
Website: \url{https://sraf.nd.edu/textual-analysis/resources/}
}
\description{
The data contains sentiments word lists.
}
\details{
The sentiment categories are: negative, positive, uncertainty, litigious, modal, and 
Harvard IV. Modal words are flagged as 1, 2 or 3, with 1 = Strong Modal, 2 = Moderate 
Modal, and 3 = Weak Modal.
}
\references{
Tim Loughran and Bill McDonald, 2011, When is a Liability not a Liability? Textual Analysis, 
Dictionaries, and 10-Ks, Journal of Finance, 66:1, 35-65.

Andriy Bodnaruk, Tim Loughran and Bill McDonald, 2015, Using 10-K Text to Gauge Financial 
Constraints, Journal of Financial and Quantitative Analysis, 50:4, 1-24.

Tim Loughran and Bill McDonald, 2016, Textual Analysis in Accounting and Finance: A Survey, 
Journal of Accounting Research, 54:4,1187-1230.
}
\keyword{data}
