% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runEDGAR_shiny.R
\name{RunEdgarShiny}
\alias{RunEdgarShiny}
\title{Shiny app for EDGAR filling management.}
\usage{
RunEdgarShiny()
}
\description{
\code{RunEdgarShiny } is a shiny app can download quarterly master index, edgar fillings,
daily master index, and do sentiment analysis on 10-K statement.
}
\details{
RunEdgarShiny  app has tabs: 'Get Master Index', 'Fillings Info', 'Download Fillings',
'Get Daily EDGAR Info', and 'Sentiment Analysis of 10K'.
User can have GUI for EDGAR filling management and for sentiment analysis
of 10-K statement. See the directory structure at home page.
Do not run the app on any browser, the app can only run on R shiny app GUI.
}
\examples{
\dontrun{

RunEdgarShiny()
}
}

