% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{cventry}
\alias{cventry}
\title{Make a Latex cventry entry}
\usage{
cventry(
  center = " ",
  lefttop = " ",
  righttop = " ",
  rightbottom = " ",
  leftbuttom = " "
)
}
\arguments{
\item{center}{Param center.}

\item{lefttop}{Param lefttop.}

\item{righttop}{Param righttop.}

\item{rightbottom}{Param rightbottom.}

\item{leftbuttom}{Param leftbuttom.}
}
\value{
A chracter vector.
}
\description{
Make a Latex cventry entry
}
\examples{
cventry(center = "Hallo", lefttop = "world")

}
