% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_wordfrquency.R
\name{GetWordfrquency}
\alias{GetWordfrquency}
\title{Creates words frequency dataframe for 10-K statement.}
\usage{
GetWordfrquency()
}
\value{
Function will return words frequency dataframe.
}
\description{
\code{GetWordfrquency} creates the word frequency dataframe for 10-K statement.
}
\details{
GetWordfrquency function ask the user to locate 10-K statement which can be downloaded using
\link[edgar]{DownloadFillings} function. Function will clean the text of 10-K statement
and creates words frequency dataframe. This words frequency dataframe will be used
in the functions \link[edgar]{PositiveWordcloud} and \link[edgar]{NegativeWordcloud}.
}
\examples{
\dontrun{

word.frq <- GetWordfrquency()
}
}

