% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordfrquency.R
\name{getWordfrquency}
\alias{getWordfrquency}
\title{Creates words frequency data frame from EDGAR filing.}
\usage{
getWordfrquency(filepath)
}
\arguments{
\item{filepath}{Path of downloaded filing.}
}
\value{
Function returns words frequency data frame.
}
\description{
\code{getWordfrquency} creates word frequency data frame using filing text.
}
\details{
getWordfrquency function takes path of filing which can be download using 
\link[edgar]{getFilings} function. Function cleans text from filing
and creates words frequency data frame. This data frame is used
in \link[edgar]{getWordcloud}, \link[edgar]{getWordHistogram}, and 
\link[edgar]{getSentimentCount} functions.
}
\examples{
\dontrun{

word.frq <- getWordfrquency('Edgar Filings/1000180_10-K_2006/1000180_10-K_2006-03-15.txt')
}
}

