% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggpie}
\alias{ggpie}
\title{Make a pie chart}
\usage{
ggpie(data, fill, legend = "none")
}
\arguments{
\item{data}{A character vector with one element.}

\item{fill}{The variable to fill the pie.}

\item{legend}{The default does not return a legend.}
}
\value{
A ggplot2 object.
}
\description{
Make a pie chart
}
\examples{
ggpie(palmerpenguins::penguins, fill = island)
}
