% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runEDGAR_shiny.R
\name{RunEdgarShiny}
\alias{RunEdgarShiny}
\title{Shiny app for EDGAR filling management.}
\usage{
RunEdgarShiny()
}
\description{
\code{RunEdgarShiny } is a shiny app which has functionality of downloading
quarterly master index, edgar filings, daily index, and do sentiment analysis
of 10-K statement.
}
\details{
RunEdgarShiny app has tabs: 'Get Master Index', 'Filings Info', 'Download Filings',
'Get Daily EDGAR Info', and 'Sentiment Analysis of 10-K'.
See the directory structure at home page. App contains pop-up interactive messages,
it can only be run on R-studio's Viewer pane.
Do not run this app on any browser.
}
\examples{
\dontrun{

RunEdgarShiny()
}
}

